refactor: add VertexAIService
